数值

+++ src/揭尹/再刷剑指offer/No6_从尾到头打印链表/Solution.java
@@ -23,7 +23,7 @@ public class Solution {
         Stack<Integer> stack = new Stack<>();
         ListNode pHead = listNode;
         while(pHead!=null){
-            stack.add((Integer) pHead.val);
+            stack.add(pHead.val);
             pHead=pHead.next;
         }
         while(!stack.isEmpty()){
